changed mainMenu.py to be able to dynamically execute any library based on FS structure
